[Frontend] Extract shared LanguageCombobox component
Consolidates the library-aware language combobox that lived inline in
both FileEditDialog and IdentifyReviewForm into a single component.
Standardizes the selected-language badge on "Name (tag)" format and
adds aria-label="Clear language" to the clear button so both dialogs
stay in sync going forward.
